[compiler] move relational IR typecheck methods out of class
